+++ lib/deck/slide.rb
@@ -93,6 +93,10 @@ module Deck
   end
 
   def <<(s)
+    if s.strip =~ /^\s*<?!VIDEO +([^\s>]*)>?$/
+      youtube_id = $1
+      s = %Q(<iframe class="video youtube" type="text/html" width="640" height="390" src="http://www.youtube.com/embed/#{youtube_id}" frameborder="0"></iframe>\n)
+    end
     @markdown_text << s
     @markdown_text << "\n"
   end